requirements.txt
djangorestframework
djoser
djangorestframework_simplejwt
django-cors-headers
django-filter
Markdown
mysqlclient